add and delete are finished working on edit

+++ item_catalog/app/record_modification/modification_controllers.py
@@ -15,7 +15,7 @@ import json
 from app import db
 
 # Import module models (i.e. User)
-from app.records.record_model import Genre, Genre, Record
+from app.records.record_model import Genre, Artist, Record
 
 # Define the blueprint: 'add', set its url prefix: app.url/
 modificationBase = Blueprint('add', __name__, url_prefix='')
@@ -25,10 +25,10 @@ modificationBase = Blueprint('add', __name__, url_prefix='')
 '''
 @modificationBase.route('/records/add', methods=['GET', 'POST'])
 def addRecords():
-#    if 'username' not in login_session:
-#        return redirect('/login')
+    if 'username' not in login_session:
+        return redirect('/login')
     if request.method == 'POST':
-        Genre_id = db.session.query(Artist.id).filter_by(artist_name = request.form['artist_Sel']).one()
+        artist_id = db.session.query(Artist.id).filter_by(artist_name = request.form['artist_Sel']).one()
         genre_id = db.session.query(Genre.id).filter_by(genre_name = request.form['genre_Sel']).one()
         newRecord = Record(
             title=request.form['title'], artist_id=int(artist_id[0]), genre_id=genre_id[0], year=int(request.form['year']), description=request.form['description'])
@@ -42,11 +42,34 @@ def addRecords():
         return render_template("records/add_records.html", artists = artists, genres = genres, loggedIn = False)
     else:
         return render_template("records/add_records.html", artists = artists, genres = genres, loggedIn = True)
+
+@modificationBase.route('/records/<int:record_id>/edit', methods=['GET', 'POST'])
+def editRecords(record_id):
+    if 'username' not in login_session:
+        return redirect('/login')
+    editRecord = db.session.query(Record.id, Record.title, Record.year, Record.description, Record.artist_id, Artist.artist_name, 
+        Genre.genre_name).filter_by(id=record_id).join(Artist).join(Genre).one()
+    artists = db.session.query(Artist).all()
+    genres = db.session.query(Genre).all()
+    if request.method == 'POST':
+        artist_id = db.session.query(Artist.id).filter_by(artist_name = request.form['artist_Sel']).one()
+        genre_id = db.session.query(Genre.id).filter_by(genre_name = request.form['genre_Sel']).one()
+        updateRecord = Record(
+            title=request.form['title'], artist_id=int(artist_id[0]), genre_id=genre_id[0], year=int(request.form['year']), description=request.form['description'])
+        db.session.add(newRecord)
+        flash('New Record Successfully Created')
+        db.session.commit()
+        return redirect('/records')
+
+    if 'username' not in login_session:
+        return render_template("/records/edit_record.html", artists = artists, genres = genres, record=editRecord, loggedIn = False)
+    else:
+        return render_template("/records/edit_record.html", artists = artists, genres = genres, record=editRecord, loggedIn = True)
     
 @modificationBase.route('/records/<int:record_id>/delete', methods=['GET', 'POST'])
 def deleteRecords(record_id):
-#    if 'username' not in login_session:
-#        return redirect('/login')
+    if 'username' not in login_session:
+        return redirect('/login')
     recordToDelete = db.session.query(Record).filter_by(id=record_id).one()
     if request.method == 'POST':
         db.session.delete(recordToDelete)
@@ -63,8 +86,8 @@ def deleteRecords(record_id):
 '''
 @modificationBase.route('/artists/add', methods=['GET', 'POST'])
 def addArtist():
-#    if 'username' not in login_session:
-#        return redirect('/login')
+    if 'username' not in login_session:
+        return redirect('/login')
     if request.method == 'POST':
         genre_id = db.session.query(Genre.id).filter_by(genre_name = request.form['genre_Sel']).one()
         newArtist = Artist(artist_name=request.form['artist_name'], genre_id=genre_id[0])
@@ -81,8 +104,8 @@ def addArtist():
     
 @modificationBase.route('/artists/<int:artist_id>/delete', methods=['GET', 'POST'])
 def deleteArtist(artist_id):
-#    if 'username' not in login_session:
-#        return redirect('/login')
+    if 'username' not in login_session:
+        return redirect('/login')
     artistToDelete = db.session.query(Artist).filter_by(id=artist_id).one()
     if request.method == 'POST':
         db.session.delete(artistToDelete)
@@ -116,8 +139,8 @@ def addGenre():
     
 @modificationBase.route('/genres/<int:genre_id>/delete', methods=['GET', 'POST'])
 def deleteGenre(genre_id):
-#    if 'username' not in login_session:
-#        return redirect('/login')
+    if 'username' not in login_session:
+        return redirect('/login')
     genreToDelete = db.session.query(Genre).filter_by(id=genre_id).one()
     if request.method == 'POST':
         db.session.delete(genreToDelete)
